requirements/base.txt
argon2-cffi==23.1.0
alembic==1.14.1
email-validator==2.3.0
fastapi==0.115.12
hydra-core==1.3.2
iopath==0.1.10
itsdangerous==2.2.0
jinja2==3.1.6
numpy==2.1.3
opencv-python-headless==4.10.0.84
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-settings==2.8.1
python-multipart==0.0.20
sqlalchemy==2.0.39
structlog==25.2.0
uvicorn[standard]==0.34.0
